backend/requirements.txt
django
djangorestframework
python-dotenv
psycopg2-binary
Django==3.0.6
Pillow==7.1.2
djangorestframework==3.12.4
django-cors-headers==3.7.0
psycopg2-binary==2.9.3
python-decouple==3.6